requirements.txt
pandas
numpy
langchain
langchain-community
sentence-transformers
faiss-cpu
ollama
streamlit
fastapi
uvicorn
python-dotenv
